chunking_service/requirements.txt
celery==5.4.0
redis==5.0.1
ffmpeg-python==0.2.0
boto3==1.35.24